Simplify API of StateMachine
Do not permit methods in the API which can request illegal state
transitions. Only operate on state + event when doing transitions, and
do not take target state as an argument. The resulting state will be
returned by the method instead, since the caller should not have to know
which state the operation should yield, it is enough that is is known
when the state machine is built.