Avoid using words like "current" and "next"
Avoid using words like "previous", "current" and "next" when talking
about states, since it will not always be clear which state it refers to
depending on the context. For example, before a state transition is
done, one state is the "current" state, but after a state transitions
has been done, the same state should be now be referred to as the
"previous" state. But this make things a bit inconsistent.

The preferred way is to use "source" and "target" state, since they are
the same both before and after a state transition has been done.